Done with the preprocess function in preprocessor.py. Included file with parameters in param.spy and created legacy folder. This branch should be merged and deleted
