+++ drumbeatid/ml_logic/preprocessor.py
@@ -1,56 +1,47 @@
 from importlib.util import spec_from_file_location
 from inspect import stack
-import wave
 
 import numpy as np
 import librosa
 
-from drumbeatid.ml_logic.functions import spectogram_mfccs, spectogram_stft, spectogram_mel
-from drumbeatid.ml_logic.functions import spectogram_chroma, padding, minmaxscaling, padding_waveforms
-
-def process_audiofile(self, audiofile):
-
-    waveform, samplingrate = librosa.load(audiofile, duration=6)
-
-    return waveform, samplingrate
+from drumbeatid.params import *
+from drumbeatid.soundprocessing.soundfeatures import Audio
+from drumbeatid.utils.minmaxscaler import CustomMinMaxScaler
 
 
-def process(waveform, samplingrate, max_length=132300):
+def preprocess(audiofile, samplingrate=22050, duration=6):
+    '''
+    Preprocessing function to prepare spectrograms as input for the model.
 
-    if waveform.shape[0] < max_length:
-        waveform = librosa.util.pad_center(waveform, size=max_length, axis=0)
+    Input:
+        audiofile: audiofile in format .wav is only implemented for the moment
+        samplingrate: samples per second to load the file using librosa
+        duration: maximum duration in seconds to take from the audiofile,
+                    default set to 6 seconds
+    '''
 
-    waveforms = np.array([waveform])
-    sr = np.array([samplingrate])
+    audio = Audio(audiofile=audiofile,  sr=samplingrate, duration=duration)
+    audio.get_all_features()
 
-    mfccs = spectogram_mfccs(waveforms, sr)
-    spectrogram_feat = spectogram_stft(waveforms)
-    melspect = spectogram_mel(waveforms, sr)
-    chroma = spectogram_chroma(waveforms, sr)
+    waveform = audio.waveform
+    sr_ = audio.samplingrate
 
-    chroma_padded = padding(chroma, size=mfccs.shape[1], axis=1)
+    audio.get_all_features()
 
-    spectogram_feat_norm = minmaxscaling(
-        spectrogram_feat, var='spectrogram', mode='transform' )
-    mfccs_norm = minmaxscaling(mfccs, var='mfccs', mode='transform')
-    melspect_norm = minmaxscaling(melspect, var='melspec', mode='transform')
-    chroma_padded_norm = minmaxscaling(chroma_padded, var='chroma', mode='transform')
+    spectrogram_list = FEATURE_LIST
+    dict_spectrogram = {}
 
-    # spectogram_feat_norm = minmaxscaling(spectrogram_feat)
-    # mfccs_norm = minmaxscaling(mfccs)
-    # melspect_norm = minmaxscaling(melspect)
-    # chroma_padded_norm = minmaxscaling(chroma_padded)
+    for feature_ in spectrogram_list:
+        scaler_ = CustomMinMaxScaler(feature=feature_)
+        scaler_.fit()
+        dict_spectrogram[feature_] = scaler_.transform(audio.features[feature_])
 
-    stacked = librosa.util.stack([mfccs_norm, melspect_norm,
-                                  chroma_padded_norm], axis=3)
-    spectogram_feat_norm = np.expand_dims(spectogram_feat_norm, axis=-1)
+    feat_stacked = librosa.util.stack([
+        dict_spectrogram['mfccs'],
+        dict_spectrogram['melspec'],
+        dict_spectrogram['chroma']
+        ], axis=2)
 
-    return spectogram_feat_norm, stacked
+    feat_spec = np.expand_dims(dict_spectrogram['spectrogram'], axis=-1)
 
-# path = '/Users/HZB/code/vickoru/drumbeatid/raw_data/groove/drummer1/session1/204_rock-halftime_140_fill_4-4.wav'
-
-# y, sr = librosa.load(path, duration=6)
-
-# X1, X2 = process(y, sr)
-# print(X1.shape)
-# print(X2.shape)
+    return feat_spec, feat_stacked
