+++ drumbeatid/params.py
@@ -1,0 +1,4 @@
+##
+
+# CONSTANTS #
+FEATURE_LIST = ['spectrogram', 'mfccs', 'chroma', 'melspec']
